Separate out tests into product-focused files
