--- Tests/StytchCoreTests/AsyncMethodsTestCase.swift
@@ -1,253 +1,0 @@
-import XCTest
-@testable import StytchCore
-
-final class AsyncMethodsTestCase: BaseTestCase {
-    func testMagicLinksEmailLoginOrCreate() async throws {
-        let container = DataContainer(data: BasicResponse(requestId: "1234", statusCode: 200))
-        let data = try Current.jsonEncoder.encode(container)
-        var request: URLRequest?
-        Current.networkingClient = .mock(verifyingRequest: { request = $0 }, returning: .success(data))
-        let baseUrl = try XCTUnwrap(URL(string: "https://myapp.com"))
-        let parameters: StytchClient.MagicLinks.Email.Parameters = .init(
-            email: "asdf@stytch.com",
-            loginMagicLinkUrl: baseUrl.appendingPathComponent("login"),
-            loginExpiration: 30,
-            signupMagicLinkUrl: baseUrl.appendingPathComponent("signup"),
-            signupExpiration: 30
-        )
-
-        XCTAssertTrue(try Current.keychainClient.get(.stytchEMLPKCECodeVerifier).isEmpty)
-
-        let response = try await StytchClient.magicLinks.email.loginOrCreate(parameters: parameters)
-        XCTAssertEqual(response.statusCode, 200)
-        XCTAssertEqual(response.requestId, "1234")
-
-        XCTAssertEqual(try Current.keychainClient.get(.stytchEMLPKCECodeVerifier), "e0683c9c02bf554ab9c731a1767bc940d71321a40fdbeac62824e7b6495a8741")
-
-        // Verify request
-        XCTAssertEqual(request?.url?.absoluteString, "https://web.stytch.com/sdk/v1/magic_links/email/login_or_create")
-        XCTAssertEqual(request?.httpMethod, "POST")
-        XCTAssertEqual(request?.httpBody, Data("{\"code_challenge_method\":\"S256\",\"signup_magic_link_url\":\"https:\\/\\/myapp.com\\/signup\",\"code_challenge\":\"V9dLhNVhiUv_9m8cwFSzLGR9l-q6NAeLskiVZ7WsjA8\",\"signup_expiration_minutes\":30,\"email\":\"asdf@stytch.com\",\"login_magic_link_url\":\"https:\\/\\/myapp.com\\/login\",\"login_expiration_minutes\":30}".utf8))
-    }
-
-    func testMagicLinksAuthenticate() async throws {
-        let authResponse: AuthenticateResponse = .mock
-        let container: DataContainer<AuthenticateResponse> = .init(data: authResponse)
-        let data = try Current.jsonEncoder.encode(container)
-        var request: URLRequest?
-        Current.networkingClient = .mock(verifyingRequest: { request = $0 }, returning: .success(data))
-        let parameters: StytchClient.MagicLinks.AuthenticateParameters = .init(
-            token: "12345",
-            sessionDuration: 15
-        )
-
-        await XCTAssertThrowsError(try await StytchClient.magicLinks.authenticate(parameters: parameters))
-
-        try Current.keychainClient.set(String.mockPKCECodeVerifier, for: .stytchEMLPKCECodeVerifier)
-
-        XCTAssertNotNil(try Current.keychainClient.get(.stytchEMLPKCECodeVerifier))
-
-        Current.timer = { _, _, _ in .init() }
-
-        let response = try await StytchClient.magicLinks.authenticate(parameters: parameters)
-        XCTAssertEqual(response.statusCode, 200)
-        XCTAssertEqual(response.requestId, "1234")
-        XCTAssertEqual(response.user.id, authResponse.user.id)
-        XCTAssertEqual(response.sessionToken, "hello_session")
-        XCTAssertEqual(response.sessionJwt, "jwt_for_me")
-        XCTAssertTrue(Calendar.current.isDate(response.session.expiresAt, equalTo: authResponse.session.expiresAt, toGranularity: .nanosecond))
-
-        XCTAssertTrue(try Current.keychainClient.get(.stytchEMLPKCECodeVerifier).isEmpty)
-
-        // Verify request
-        XCTAssertEqual(request?.url?.absoluteString, "https://web.stytch.com/sdk/v1/magic_links/authenticate")
-        XCTAssertEqual(request?.httpMethod, "POST")
-        XCTAssertEqual(request?.httpBody, Data("{\"token\":\"12345\",\"session_duration_minutes\":15,\"code_verifier\":\"e0683c9c02bf554ab9c731a1767bc940d71321a40fdbeac62824e7b6495a8741\"}".utf8))
-    }
-
-    func testSessionsAuthenticate() async throws {
-        let authResponse: AuthenticateResponse = .mock
-        let container: DataContainer<AuthenticateResponse> = .init(data: authResponse)
-        let data = try Current.jsonEncoder.encode(container)
-        var request: URLRequest?
-        Current.networkingClient = .mock(verifyingRequest: { request = $0 }, returning: .success(data))
-
-        let parameters: StytchClient.Sessions.AuthenticateParameters = .init(sessionDuration: 15)
-
-        Current.timer = { _, _, _ in .init() }
-
-        Current.sessionStorage.updateSession(
-            .mock(userId: "i_am_user"),
-            tokens: [.jwt("i'm_jwt"), .opaque("opaque_all_day")],
-            hostUrl: try XCTUnwrap(URL(string: "https://url.com"))
-        )
-
-        let response = try await StytchClient.sessions.authenticate(parameters: parameters)
-
-        XCTAssertEqual(response.user.id, authResponse.user.id)
-        XCTAssertEqual(response.sessionToken, "hello_session")
-        XCTAssertEqual(response.sessionJwt, "jwt_for_me")
-        XCTAssertTrue(Calendar.current.isDate(response.session.expiresAt, equalTo: authResponse.session.expiresAt, toGranularity: .nanosecond))
-
-        // Verify request
-        XCTAssertEqual(request?.url?.absoluteString, "https://web.stytch.com/sdk/v1/sessions/authenticate")
-        XCTAssertEqual(request?.httpMethod, "POST")
-        XCTAssertEqual(request?.httpBody, Data("{\"session_duration_minutes\":15,\"session_token\":\"opaque_all_day\"}".utf8))
-
-        XCTAssertEqual(StytchClient.sessions.sessionJwt, .jwt("jwt_for_me"))
-        XCTAssertEqual(StytchClient.sessions.sessionToken, .opaque("hello_session"))
-    }
-
-    func testSessionsRevoke() async throws {
-        let container: DataContainer<BasicResponse> = .init(data: .init(requestId: "request_id", statusCode: 200))
-        let data = try Current.jsonEncoder.encode(container)
-        var request: URLRequest?
-        Current.networkingClient = .mock(verifyingRequest: { request = $0 }, returning: .success(data))
-
-        Current.timer = { _, _, _ in .init() }
-
-        Current.sessionStorage.updateSession(
-            .mock(userId: "i_am_user"),
-            tokens: [.jwt("i'm_jwt"), .opaque("opaque_all_day")],
-            hostUrl: try XCTUnwrap(URL(string: "https://url.com"))
-        )
-
-        XCTAssertEqual(StytchClient.sessions.sessionToken, .opaque("opaque_all_day"))
-        XCTAssertEqual(StytchClient.sessions.sessionJwt, .jwt("i'm_jwt"))
-
-        let response = try await StytchClient.sessions.revoke()
-        XCTAssertEqual(response.statusCode, 200)
-        XCTAssertEqual(response.requestId, "request_id")
-
-        // Verify request
-        XCTAssertEqual(request?.url?.absoluteString, "https://web.stytch.com/sdk/v1/sessions/revoke")
-        XCTAssertEqual(request?.httpMethod, "POST")
-        XCTAssertEqual(request?.httpBody, Data("{\"session_token\":\"opaque_all_day\"}".utf8))
-
-        XCTAssertNil(StytchClient.sessions.sessionJwt)
-        XCTAssertNil(StytchClient.sessions.sessionToken)
-    }
-
-    func testOtpLoginOrCreate() async throws {
-        let container: DataContainer<StytchClient.OneTimePasscodes.LoginOrCreateResponse> = .init(
-            data: .init(
-                requestId: "1234",
-                statusCode: 200,
-                wrapped: .init(methodId: "method_id_1234")
-            )
-        )
-        let data = try Current.jsonEncoder.encode(container)
-        var request: URLRequest?
-        Current.networkingClient = .mock(verifyingRequest: { request = $0 }, returning: .success(data), .success(data), .success(data))
-
-        try await [
-            (
-                StytchClient.OneTimePasscodes.LoginOrCreateParameters(deliveryMethod: .whatsapp(phoneNumber: "+12345678901"), expiration: 3),
-                "https://web.stytch.com/sdk/v1/otps/whatsapp/login_or_create",
-                "{\"expiration_minutes\":3,\"phone_number\":\"+12345678901\"}"
-            ),
-            (
-                .init(deliveryMethod: .sms(phoneNumber: "+11098765432")),
-                "https://web.stytch.com/sdk/v1/otps/sms/login_or_create",
-                "{\"phone_number\":\"+11098765432\"}"
-            ),
-            (
-                .init(deliveryMethod: .email("test@stytch.com")),
-                "https://web.stytch.com/sdk/v1/otps/email/login_or_create",
-                "{\"email\":\"test@stytch.com\"}"
-            ),
-        ]
-        .asyncForEach { params, urlString, body in
-            let response = try await StytchClient.otps.loginOrCreate(parameters: params)
-            XCTAssertEqual(response.methodId, "method_id_1234")
-            XCTAssertEqual(response.statusCode, 200)
-            XCTAssertEqual(response.requestId, "1234")
-
-            // Verify request
-            guard let request = request else { XCTFail("Request should be present"); return }
-
-            XCTAssertEqual(request.url?.absoluteString, urlString)
-            XCTAssertEqual(request.httpMethod, "POST")
-            XCTAssertEqual(request.httpBody, Data(body.utf8))
-
-            XCTAssertNil(StytchClient.sessions.sessionJwt)
-            XCTAssertNil(StytchClient.sessions.sessionToken)
-        }
-    }
-
-    func testOtpAuthenticate() async throws {
-        let authResponse: AuthenticateResponse = .mock
-        let container: DataContainer<AuthenticateResponse> = .init(data: authResponse)
-        let data = try Current.jsonEncoder.encode(container)
-        var request: URLRequest?
-        Current.networkingClient = .mock(verifyingRequest: { request = $0 }, returning: .success(data))
-        let parameters: StytchClient.OneTimePasscodes.AuthenticateParameters = .init(
-            code: "i_am_code",
-            methodId: "method_id_fake_id",
-            sessionDuration: 20
-        )
-
-        XCTAssertNil(StytchClient.sessions.sessionToken)
-        XCTAssertNil(StytchClient.sessions.sessionJwt)
-
-        Current.timer = { _, _, _ in .init() }
-
-        let response = try await StytchClient.otps.authenticate(parameters: parameters)
-        XCTAssertEqual(response.statusCode, 200)
-        XCTAssertEqual(response.requestId, "1234")
-        XCTAssertEqual(response.user.id, authResponse.user.id)
-        XCTAssertEqual(response.sessionToken, "hello_session")
-        XCTAssertEqual(response.sessionJwt, "jwt_for_me")
-        XCTAssertTrue(Calendar.current.isDate(response.session.expiresAt, equalTo: authResponse.session.expiresAt, toGranularity: .nanosecond))
-
-        XCTAssertEqual(StytchClient.sessions.sessionToken, .opaque("hello_session"))
-        XCTAssertEqual(StytchClient.sessions.sessionJwt, .jwt("jwt_for_me"))
-
-        // Verify request
-        XCTAssertEqual(request?.url?.absoluteString, "https://web.stytch.com/sdk/v1/otps/authenticate")
-        XCTAssertEqual(request?.httpMethod, "POST")
-        XCTAssertEqual(request?.httpBody, Data("{\"token\":\"i_am_code\",\"method_id\":\"method_id_fake_id\",\"session_duration_minutes\":20}".utf8))
-    }
-
-    func testHandleUrl() async throws {
-        let authResponse: AuthenticateResponse = .mock
-        let container: DataContainer<AuthenticateResponse> = .init(data: authResponse)
-        let data = try Current.jsonEncoder.encode(container)
-        Current.networkingClient = .mock(returning: .success(data))
-
-        let notHandledUrl = try XCTUnwrap(URL(string: "https://myapp.com?token=12345"))
-
-        switch try await StytchClient.handle(url: notHandledUrl, sessionDuration: 30) {
-        case .handled:
-            XCTFail("expected to be nothandled")
-        case .notHandled, .manualHandlingRequired:
-            break
-        }
-
-        let handledUrl = try XCTUnwrap(URL(string: "https://myapp.com?token=12345&stytch_token_type=magic_links"))
-
-        await XCTAssertThrowsError(try await StytchClient.handle(url: handledUrl))
-
-        try Current.keychainClient.set(String.mockPKCECodeVerifier, for: .stytchEMLPKCECodeVerifier)
-
-        Current.timer = { _, _, _ in .init() }
-
-        switch try await StytchClient.handle(url: handledUrl, sessionDuration: 30) {
-        case let .handled(response):
-            XCTAssertEqual(response.sessionJwt, "jwt_for_me")
-            XCTAssertEqual(response.session.authenticationFactors.count, 1)
-        case .notHandled, .manualHandlingRequired:
-            XCTFail("expected to be handled")
-        }
-    }
-
-    func testExternalSessionsUpdate() {
-        XCTAssertNil(StytchClient.sessions.sessionToken)
-        XCTAssertNil(StytchClient.sessions.sessionJwt)
-
-        StytchClient.sessions.update(sessionTokens: [.opaque("token"), .jwt("jwt")])
-
-        XCTAssertEqual(StytchClient.sessions.sessionToken, .opaque("token"))
-        XCTAssertEqual(StytchClient.sessions.sessionJwt, .jwt("jwt"))
-    }
-}
